reorganized file and folder structure. created navbar.js create app.js and some opther stuff idk
