+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/williamjacobs/bootcamp/project_3_folder/ZTC_TodoApp/client/src/components/signUp.js":"1","/Users/williamjacobs/bootcamp/project_3_folder/ZTC_TodoApp/client/src/index.js":"2"},{"size":223,"mtime":1607200642363,"results":"3","hashOfConfig":"4"},{"size":599,"mtime":1607201162001,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"16vdflj",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/williamjacobs/bootcamp/project_3_folder/ZTC_TodoApp/client/src/components/signUp.js",[],"/Users/williamjacobs/bootcamp/project_3_folder/ZTC_TodoApp/client/src/index.js",[]]
+[{"/Users/williamjacobs/bootcamp/project_3_folder/ZTC_TodoApp/client/src/index.js":"1","/Users/williamjacobs/bootcamp/project_3_folder/ZTC_TodoApp/client/src/components/Navbar.js":"2","/Users/williamjacobs/bootcamp/project_3_folder/ZTC_TodoApp/client/src/App.js":"3"},{"size":613,"mtime":1607204082198,"results":"4","hashOfConfig":"5"},{"size":993,"mtime":1607204102482,"results":"6","hashOfConfig":"5"},{"size":339,"mtime":1607204103418,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"16vdflj",{"filePath":"10","messages":"11","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/williamjacobs/bootcamp/project_3_folder/ZTC_TodoApp/client/src/index.js",[],"/Users/williamjacobs/bootcamp/project_3_folder/ZTC_TodoApp/client/src/components/Navbar.js",["14"],"/Users/williamjacobs/bootcamp/project_3_folder/ZTC_TodoApp/client/src/App.js",[],{"ruleId":"15","severity":1,"message":"16","line":2,"column":17,"nodeType":"17","messageId":"18","endLine":2,"endColumn":26},"no-unused-vars","'Component' is defined but never used.","Identifier","unusedVar"]
